assistant_comptable_api/requirements.txt
django
ipython
# google-generativeai
django-browser-reload
django-extensions
django-import-export
django-rest-framework
pillow
pypdf
# paypalrestsdk
python-decouple
django-paypal
#pip install -i https://test.pypi.org/simple/ cinetpay-sdk==0.1.1